+++ packages/pkg/test/basic.spec.ts
@@ -1,5 +1,0 @@
-import echo from 'pkg'
-
-test('it should just work', () => {
-  expect(echo()).toBe('Hello World!')
-})