+++ packages/preset-node-lib/shim.d.ts
@@ -1,0 +1,7 @@
+declare module '@size-limit/file' {
+  export = [
+    {
+      name: '@size-limit/file',
+    },
+  ]
+}