+++ packages/preset-node-lib/src/index.ts
@@ -1,0 +1,6 @@
+/// <reference path="../shim.d.ts" />
+
+import file from '@size-limit/file'
+import nodeEsbuild from 'size-limit-node-esbuild'
+
+export default [...nodeEsbuild, ...file] as const