+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-- [pkg](./packages/pkg/CHANGELOG.md)
+- [size-limit-node-esbuild](./packages/node-esbuild/CHANGELOG.md)
+- [size-limit-preset-node-lib](./packages/preset-node-lib/CHANGELOG.md)